requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annoy==1.17.3
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
bitarray==3.5.0
bleach==6.2.0
bokeh==3.7.3
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dask==2025.7.0
datasets==4.0.0
debugpy==1.8.14
decorator==5.2.1
diffusers==0.34.0
dill==0.3.8
executing==2.2.0
faiss-cpu==1.11.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.5
holoviews==1.21.0
huggingface-hub==0.33.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_bokeh==4.0.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kaggle==1.7.4.5
kiwisolver==1.4.8
lazy_loader==0.4
linkify-it-py==2.0.3
locket==1.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
narwhals==1.48.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pandas==2.3.1
panel==1.7.4
param==2.2.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
simhash==2.1.2
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.2
toolz==1.0.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.14
xxhash==3.5.0
xyzservices==2025.4.0
yarl==1.20.1
zipp==3.23.0
